[TASK] Move cache-related functions to `CacheService`
